+++ resources/ext.layers.editor/CanvasManager.js
@@ -36,7 +36,6 @@
 		TextInputController: 'Canvas.TextInputController',
 		AlignmentController: 'Canvas.AlignmentController',
 		SmartGuidesController: 'Canvas.SmartGuidesController',
-		EyedropperController: 'Canvas.EyedropperController',
 		RenderCoordinator: 'Canvas.RenderCoordinator',
 		InteractionController: 'Canvas.InteractionController',
 		ValidationManager: 'Validation.Manager',
@@ -252,8 +251,7 @@ class CanvasManager {
 			[ 'ClipboardController', 'clipboardController' ],
 			[ 'TextInputController', 'textInputController' ],
 			[ 'AlignmentController', 'alignmentController' ],
-			[ 'SmartGuidesController', 'smartGuidesController' ],
-			[ 'EyedropperController', 'eyedropperController' ]
+			[ 'SmartGuidesController', 'smartGuidesController' ]
 		];
 
 		controllers.forEach( ( entry ) => {
@@ -1811,7 +1809,6 @@ class CanvasManager {
 			'drawingController',
 			'clipboardController',
 			'smartGuidesController',
-			'eyedropperController',
 			'styleController'
 		];